Implement caching of get all user ids in redis (#10)
