+++ .env
@@ -8,6 +8,7 @@ ACTIX_DEMO_HASH_COST     = 8
 ACTIX_DEMO_LOGGER_FORMAT = pretty
 ACTIX_DEMO_JWT_KEY       = test
 ACTIX_DEMO_REDIS_URL     = redis://127.0.0.1
+CACHED_REDIS_CONNECTION_STRING = redis://127.0.0.1
 ACTIX_DEMO_JOB_BIN_PATH  = /bin/echo
 ACTIX_DEMO_RATE_LIMIT_AUTH_MAX_REQUESTS  = 5
 ACTIX_DEMO_RATE_LIMIT_AUTH_WINDOW_SECS   = 120
