+++ src/actions/users.rs
@@ -1,7 +1,4 @@
 use diesel::prelude::*;
-use diesel::r2d2::ConnectionManager;
-use diesel_tracing::pg::InstrumentedPgConnection;
-use r2d2::PooledConnection;
 
 use crate::errors::DomainError;
 use crate::models::misc::Pagination;
@@ -10,13 +7,16 @@ use crate::models::users::{
     NewUser, Password, User, UserAuthDetails, UserAuthDetailsWithRoles, UserId,
     UserWithRoles, Username,
 };
+use crate::types::DbConnection;
 use bcrypt::hash;
+use cached::proc_macro::io_cached;
+use cached::{IOCached, RedisCache};
 use do_notation::m;
 use validators::prelude::*;
 
 pub fn get_roles_for_user(
     uid: &UserId,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Vec<RoleEnum>, DomainError> {
     use crate::schema::roles::dsl as roles;
     use crate::schema::users_roles::dsl as users_roles;
@@ -29,7 +29,7 @@ pub fn get_roles_for_user(
 
 pub fn get_roles_for_users(
     users: Vec<User>,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Vec<UserWithRoles>, DomainError> {
     users
         .into_iter()
@@ -42,7 +42,7 @@ pub fn get_roles_for_users(
 
 pub fn find_user_by_uid(
     uid: &UserId,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Option<UserWithRoles>, DomainError> {
     use crate::schema::users::dsl as users;
 
@@ -61,7 +61,7 @@ pub fn find_user_by_uid(
 
 pub fn find_user_by_name(
     user_name: &Username,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Option<UserWithRoles>, DomainError> {
     use crate::schema::users::dsl as users;
 
@@ -94,7 +94,7 @@ pub fn find_user_by_name(
 
 pub fn get_user_auth_details(
     user_name: &Username,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Option<UserAuthDetailsWithRoles>, DomainError> {
     use crate::schema::users::dsl as users;
 
@@ -122,7 +122,7 @@ pub fn get_user_auth_details(
 
 pub fn get_all_users(
     pagination: &Pagination,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Vec<UserWithRoles>, DomainError> {
     use crate::schema::users::dsl as users;
 
@@ -138,8 +138,30 @@ pub fn get_all_users(
     })
 }
 
+fn build_user_ids_cache() -> RedisCache<String, Vec<UserId>> {
+    RedisCache::new("user_ids", 3600)
+        .build()
+        .map_err(|e| {
+            DomainError::new_internal_error(format!(
+                "Failed to build cache: {e:?}"
+            ))
+        })
+        .expect("Failed to create Redis cache")
+}
+
+lazy_static::lazy_static! {
+    static ref USER_IDS_CACHE: RedisCache<String, Vec<UserId>> =
+        build_user_ids_cache();
+}
+
+#[io_cached(
+    map_error = r##"|e| DomainError::new_internal_error(format!("Redis error: {:?}", e))"##,
+    ty = "RedisCache<String, Vec<UserId>>",
+    create = r##" { build_user_ids_cache() } "##, 
+    convert = r#"{ "user_id".to_string() }"#  // Static key for the entire user list
+)]
 pub fn get_all_user_ids(
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Vec<UserId>, DomainError> {
     use crate::schema::users::dsl as users;
 
@@ -154,7 +176,7 @@ pub fn get_all_user_ids(
 pub fn search_users(
     query: &str,
     pagination: &Pagination,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<Vec<UserWithRoles>, DomainError> {
     use crate::schema::users::dsl as users;
 
@@ -175,7 +197,7 @@ pub fn insert_new_user(
     nu: NewUser,
     role: RoleEnum,
     hash_cost: u32,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<UserWithRoles, DomainError> {
     use crate::schema::roles::dsl as roles;
     use crate::schema::users::dsl as users;
@@ -189,6 +211,7 @@ pub fn insert_new_user(
         })?;
         nu2
     };
+
     conn.transaction(|conn| {
         let _ = diesel::insert_into(users::users)
             .values(&nu)
@@ -211,19 +234,26 @@ pub fn insert_new_user(
 
         let roles = get_roles_for_user(&user.id, conn)?;
 
-        Ok(UserWithRoles {
+        let user_with_roles = UserWithRoles {
             id: user.id,
             username: user.username,
             created_at: user.created_at,
             roles,
-        })
+        };
+
+        // Invalidate the cache since we've added a new user
+        if let Err(e) = &USER_IDS_CACHE.cache_remove(&"user_id".to_owned()) {
+            tracing::error!(error = %e, "Failed to invalidate user IDs cache");
+        }
+
+        Ok(user_with_roles)
     })
 }
 
 pub fn insert_new_regular_user(
     nu: NewUser,
     hash_cost: u32,
-    conn: &mut PooledConnection<ConnectionManager<InstrumentedPgConnection>>,
+    conn: &mut DbConnection,
 ) -> Result<UserWithRoles, DomainError> {
     insert_new_user(nu, RoleEnum::RoleUser, hash_cost, conn)
 }
